refactor(tooling): pay off the stale-suppression debt and stub ext-pcntl
Psalm reported 150 @psalm-suppress annotations that no longer suppressed
anything. Removed precisely rather than line-wise: a line may carry several
issue names where only some have gone stale, and deleting the whole line would
silently re-enable a suppression still doing work. Psalm names the dead entry,
so only that name was taken.

Two of them were hiding real defects.

ConsoleTailCommand::execute() declared `: int` and polled in an unconditional
for(;;) behind a suppression saying so. A command that can only end by killing
the process returns no exit code to the shell, runs no `finally`, and hangs any
test that touches it. It now installs SIGINT/SIGTERM handlers where the platform
supports them and unwinds normally.

Fifteen `(int) SIG*` casts existed only because the analysers could not see that
SIGINT and friends are integers. ext-pcntl is POSIX-only — PHP ships no Windows
build of it — so on a developer machine every signal constant read as undefined
and every pcntl_*() argument as mixed, in the worker-lifecycle code where signals
actually matter. phpstorm-stubs carries pcntl.php and pcntl_c.php; wiring both in
closed the gap, and the casts it had provoked are gone. That single change
surfaced 29 further findings elsewhere, which is the measure of how blind the
analysis had been there.

findUnusedPsalmSuppress is now on permanently: the debt it named is paid, so it
guards from here. The two array-offset switches stay off with their exact counts
recorded — 267 and 219 — because paying those means giving 486 array accesses a
proved-present offset, and that is its own campaign.

Also: .gitattributes pins the working tree to LF, not just the stored blobs.
"LF will be replaced by CRLF" was the conversion announcing itself on every git
command, and the fix is to remove the conversion, not reverse it. CRLF was the
tempting answer and would have been silent corruption: 364 PHP files carry
heredocs, and a heredoc takes the file's own line endings into the string.

+++ extensions/mcp-server/src/Internal/Subprocess/SubprocessRunner.php
@@ -275,8 +275,7 @@ final class SubprocessRunner
         if (PHP_OS_FAMILY === 'Windows' && $pid > 0) {
             $this->windowsForceKillTree($pid);
         } else {
-            /** @psalm-suppress UndefinedConstant SIGKILL is POSIX-only, guarded via defined() */
-            $signal = defined('SIGKILL') ? (int) SIGKILL : 9;
+            $signal = defined('SIGKILL') ? SIGKILL : 9;
             proc_terminate($process, $signal);
         }
