+++ src/Runtime/FrankenPhpRuntime.php
@@ -292,12 +292,9 @@ final class FrankenPhpRuntime implements ReloadableRuntimeInterface, SupportsEar
             return;
         }
 
-        /** @psalm-suppress UndefinedConstant POSIX-only, guarded by OS check */
-        $sigint = (int) SIGINT;
-        /** @psalm-suppress UndefinedConstant */
-        $sigterm = (int) SIGTERM;
-        /** @psalm-suppress UndefinedConstant */
-        $sigusr1 = (int) SIGUSR1;
+        $sigint = SIGINT;
+        $sigterm = SIGTERM;
+        $sigusr1 = SIGUSR1;
 
         pcntl_signal($sigint, function (): void {
             $this->stop();
